+++ crates/cbf-chrome/src/lib.rs
@@ -4,10 +4,12 @@
 //! integration while keeping `cbf` browser-generic.
 
 pub mod backend;
-pub mod browser;
+mod browser;
 pub mod command;
 pub mod data;
 pub mod event;
 pub mod ffi;
 pub mod platform;
 pub mod process;
+
+pub use browser::ChromiumBrowserHandleExt;